+++ src/app/[eventId]/_components/EventDetails.tsx
