+++ src/app/[eventId]/_components/ParticipantsNav.tsx
