+++ src/app/[eventId]/_components/SearchParticipants.tsx
