+++ src/app/[eventId]/_components/participants/ParticipantCard.tsx
