+++ src/app/[eventId]/_components/participants/Participants.tsx
