+++ src/app/layout.tsx
@@ -2,7 +2,7 @@ import type { Metadata } from 'next';
 import localFont from 'next/font/local';
 import '@/styles/globals.css';
 import ReactQueryProviders from '@/lib/provider/ReactQueryProviders';
-import Background from './components/Background';
+import Background from './_components/Background';
 
 const consola = localFont({
   src: '../styles/font/Satoshi-Variable.ttf',
